requirements.txt
flask
python-dotenv
google-genai
PyPDF2
gunicorn
